backend/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
grpcio==1.62.2
idna==3.7
oauthlib==3.2.2
protobuf==5.26.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
tweepy @ git+https://github.com/tweepy/tweepy.git@91a41c6e1c955d278c370d51d5cf43b05f7cd979
urllib3==2.2.1
xai-sdk==0.0.4
